IMPORTANT!!! Gitignore deleting stuff weirdness
So I inadvertently stashed stuff (switched branches with an uncommitted
change in the GUI) which apparently runs `gitclean -d`, which removes
untracked files/folders. I had not ignored the data folder but had
ignored the extensions of all of the files in it (*.json and *.csv). My
data folder was gone! And since the files weren't tracked I couldn't
roll back to it. Fortunately there was nothing irreplacable in there
but there could have been. <a
href="http://blog.icefusion.co.uk/git-stash-can-delete-ignored-files-git
-stash-u/">This post</a> says that /directory/* gets you in trouble
while /directory is safe. My "versions" directory stayed put, which was
ignored like this: `versions/'. I'm changing it to the other format
just to be safe.